# fix: Optimize span and trace ID generation