This improves the performance of span and trace ID generation by avoiding the loop overhead in the common case of `rand` returning a valid ID.

Benchmark results:
```
Comparison:
generate_trace_id_while:  8023027.6 i/s
   generate_trace_id:  4328017.1 i/s - 1.85x slower
...
Comparison:
generate_span_id_while:  8364182.8 i/s
    generate_span_id:  4382207.1 i/s - 1.91x slower
```